fix compile error

+++ src/wrapper/map.rs
@@ -12,7 +12,7 @@ pub fn get_map_value(env: NIF_ENV, map: NIF_TERM, key: NIF_TERM) -> Option<NIF_T
     Some(result)
 }
 
-pub fn get_map_size(env: NIF_ENV, map: NIF_TERM) -> Option<u64> {
+pub fn get_map_size(env: NIF_ENV, map: NIF_TERM) -> Option<usize> {
     let mut size: nif_interface::size_t = unsafe { mem::uninitialized() };
     let success = unsafe { nif_interface::enif_get_map_size(env, map, &mut size as *mut nif_interface::size_t) };
